Clean up DefaultResourceService
